# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lendora short-interest API",
    "version": "1.0.0",
    "description": "Live short interest for Lendora markets on Robinhood Chain (docs/prd/07). Scope: Lendora markets only.\n\nEvery data response carries `asOfBlock`, `asOfTime` and `confirmed` (finalized) — SI-R13.\n\nLimits (SI-R10): free 60 req/min and 1 WebSocket connection per IP; keyed 600 req/min and 10 connections. Create a key with Sign-In with Ethereum at `POST /v1/auth/keys`.\n\nWebSocket `/v1/stream`: send `{\"channel\":\"market\",\"symbol\":\"NVDA\"}` or `{\"channel\":\"events\",\"symbol\":\"*\"}`; `\"op\":\"unsubscribe\"` to stop.\n\nChain id 46630. Data as is, no warranty; attribution requested (/v1/terms).",
    "license": {
      "name": "Data terms",
      "url": "https://lendora.xyz/terms/data"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
      "MarketsResponse": {
        "type": "object",
        "properties": {
          "asOfBlock": {
            "type": "string",
            "pattern": "^-?\\d+$",
            "description": "Block the data is as of (SI-R13)",
            "example": "1000000000000000000"
          },
          "asOfTime": {
            "type": "string",
            "description": "ISO 8601 UTC",
            "example": "2026-10-01T16:00:00.000Z"
          },
          "confirmed": {
            "type": "boolean",
            "description": "asOfBlock ≤ the chain's finalized block (SI-R3)"
          },
          "safe": {
            "type": "boolean",
            "description": "asOfBlock ≤ the chain's safe block"
          },
          "scope": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Market"
            }
          }
        },
        "required": [
          "asOfBlock",
          "asOfTime",
          "confirmed",
          "safe",
          "scope",
          "data"
        ]
      },
...
    },
...
  },
  "paths": {
    "/v1/markets": {
      "get": {
        "tags": [
          "markets"
        ],
        "summary": "All stocks, current snapshot (07 §Definitions)",
        "responses": {
          "200": {
            "description": "Snapshots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (SI-R10)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Indexer not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{symbol}": {
      "get": {
        "tags": [
          "markets"
        ],
        "summary": "One stock: snapshot + params (LLTV, caps, U_MAX, buffer now, next session)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9.]{1,12}$",
              "example": "NVDA"
            },
            "required": true,
            "name": "symbol",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot and parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (SI-R10)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Indexer not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{symbol}/history": {
      "get": {
        "tags": [
          "markets"
        ],
        "summary": "Time series of the snapshot fields (1m / 1h / 1d buckets); `format=csv` for CSV (SI-R12)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9.]{1,12}$",
              "example": "NVDA"
            },
            "required": true,
            "name": "symbol",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "1m",
                "1h",
                "1d"
              ],
              "default": "1h"
            },
            "required": false,
            "name": "interval",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}(T\\d{2}:\\d{2}(:\\d{2}(\\.\\d{1,9})?)?(Z|[+-]\\d{2}:\\d{2})?)?$",
              "description": "Inclusive; default: 7 days before `to`",
              "example": "2026-10-01T16:00:00.000Z"
            },
            "required": false,
            "description": "Inclusive; default: 7 days before `to`",
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}(T\\d{2}:\\d{2}(:\\d{2}(\\.\\d{1,9})?)?(Z|[+-]\\d{2}:\\d{2})?)?$",
              "description": "Exclusive; default: now",
              "example": "2026-10-01T16:00:00.000Z"
            },
            "required": false,
            "description": "Exclusive; default: now",
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ],
              "default": "json",
              "description": "SI-R12"
            },
            "required": false,
            "description": "SI-R12",
            "name": "format",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5000,
              "default": 2000
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "History",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (SI-R10)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Indexer not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/positions/{address}": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "Positions of a wallet in Lendora markets (public onchain data)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$"
            },
            "required": true,
            "name": "address",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Positions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (SI-R10)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Indexer not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/status": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Oracle freshness, guard state and indexer lag per market",
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (SI-R10)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Indexer not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}